+++ spring-integration-core/src/main/java/org/springframework/integration/config/xml/GatewayParser.java
@@ -23,20 +23,33 @@ import java.util.Map;
 
 import org.w3c.dom.Element;
 
+import org.springframework.beans.factory.BeanDefinitionStoreException;
+import org.springframework.beans.factory.annotation.AnnotatedGenericBeanDefinition;
 import org.springframework.beans.factory.config.BeanDefinition;
 import org.springframework.beans.factory.config.BeanDefinitionHolder;
+import org.springframework.beans.factory.config.ConfigurableBeanFactory;
 import org.springframework.beans.factory.support.BeanDefinitionBuilder;
 import org.springframework.beans.factory.support.BeanDefinitionReaderUtils;
+import org.springframework.beans.factory.support.BeanDefinitionRegistry;
+import org.springframework.beans.factory.support.BeanNameGenerator;
 import org.springframework.beans.factory.support.ManagedMap;
+import org.springframework.beans.factory.support.RootBeanDefinition;
 import org.springframework.beans.factory.xml.AbstractBeanDefinitionParser;
 import org.springframework.beans.factory.xml.BeanDefinitionParser;
 import org.springframework.beans.factory.xml.ParserContext;
+import org.springframework.context.ConfigurableApplicationContext;
+import org.springframework.core.ResolvableType;
+import org.springframework.expression.common.LiteralExpression;
+import org.springframework.integration.annotation.AnnotationConstants;
 import org.springframework.integration.config.ExpressionFactoryBean;
 import org.springframework.integration.config.IntegrationConfigUtils;
-import org.springframework.integration.config.MessagingGatewayRegistrar;
 import org.springframework.integration.gateway.GatewayMethodMetadata;
+import org.springframework.integration.gateway.GatewayProxyFactoryBean;
+import org.springframework.integration.gateway.RequestReplyExchanger;
 import org.springframework.util.Assert;
+import org.springframework.util.ClassUtils;
 import org.springframework.util.CollectionUtils;
+import org.springframework.util.ObjectUtils;
 import org.springframework.util.StringUtils;
 import org.springframework.util.xml.DomUtils;
 
@@ -50,7 +63,7 @@ import org.springframework.util.xml.DomUtils;
  */
 public class GatewayParser implements BeanDefinitionParser {
 
-	private final MessagingGatewayRegistrar registrar = new MessagingGatewayRegistrar();
+	private static final String PROXY_DEFAULT_METHODS_ATTR = "proxyDefaultMethods";
 
 	@Override
 	public BeanDefinition parse(final Element element, ParserContext parserContext) {
@@ -89,8 +102,7 @@ public class GatewayParser implements BeanDefinitionParser {
 
 		gatewayAttributes.put("proxyDefaultMethods", element.getAttribute("proxy-default-methods"));
 
-		BeanDefinitionHolder gatewayHolder = this.registrar.parse(gatewayAttributes, parserContext.getRegistry(),
-				IntegrationConfigUtils.annotationBeanNameGenerator(parserContext.getRegistry()));
+		BeanDefinitionHolder gatewayHolder = buildBeanDefinition(gatewayAttributes, parserContext);
 		if (isNested) {
 			return gatewayHolder.getBeanDefinition();
 		}
@@ -100,7 +112,7 @@ public class GatewayParser implements BeanDefinitionParser {
 		}
 	}
 
-	private void headers(Element element, Map<String, Object> gatewayAttributes) {
+	private static void headers(Element element, Map<String, Object> gatewayAttributes) {
 		List<Element> headerElements = DomUtils.getChildElementsByTagName(element, "default-header");
 		if (!CollectionUtils.isEmpty(headerElements)) {
 			List<Map<String, Object>> headers = new ArrayList<>(headerElements.size());
@@ -116,8 +128,9 @@ public class GatewayParser implements BeanDefinitionParser {
 		}
 	}
 
-	private void methods(final Element element, ParserContext parserContext,
+	private static void methods(Element element, ParserContext parserContext,
 			final Map<String, Object> gatewayAttributes) {
+
 		List<Element> methodElements = DomUtils.getChildElementsByTagName(element, "method");
 		if (!CollectionUtils.isEmpty(methodElements)) {
 			Map<String, BeanDefinition> methodMetadataMap = new ManagedMap<>();
@@ -162,6 +175,148 @@ public class GatewayParser implements BeanDefinitionParser {
 			}
 
 			gatewayAttributes.put("methods", methodMetadataMap);
+		}
+	}
+
+
+	private static BeanDefinitionHolder buildBeanDefinition(Map<String, Object> gatewayAttributes, // NOSONAR complexity
+			ParserContext parserContext) {
+
+		BeanDefinitionRegistry registry = parserContext.getRegistry();
+		String defaultPayloadExpression = (String) gatewayAttributes.get("defaultPayloadExpression");
+
+		@SuppressWarnings("unchecked")
+		Map<String, Object>[] defaultHeaders = (Map<String, Object>[]) gatewayAttributes.get("defaultHeaders");
+
+		String defaultRequestChannel = (String) gatewayAttributes.get("defaultRequestChannel");
+		String defaultReplyChannel = (String) gatewayAttributes.get("defaultReplyChannel");
+		String errorChannel = (String) gatewayAttributes.get("errorChannel");
+		String asyncExecutor = (String) gatewayAttributes.get("asyncExecutor");
+		String mapper = (String) gatewayAttributes.get("mapper");
+		String proxyDefaultMethods = (String) gatewayAttributes.get(PROXY_DEFAULT_METHODS_ATTR);
+
+		boolean hasMapper = StringUtils.hasText(mapper);
+		boolean hasDefaultPayloadExpression = StringUtils.hasText(defaultPayloadExpression);
+		Assert.state(!hasMapper || !hasDefaultPayloadExpression,
+				"'defaultPayloadExpression' is not allowed when a 'mapper' is provided");
+
+		boolean hasDefaultHeaders = !ObjectUtils.isEmpty(defaultHeaders);
+		Assert.state(!hasMapper || !hasDefaultHeaders,
+				"'defaultHeaders' are not allowed when a 'mapper' is provided");
+
+		ConfigurableBeanFactory beanFactory = obtainBeanFactory(registry);
+		Class<?> serviceInterface = getServiceInterface((String) gatewayAttributes.get("serviceInterface"), beanFactory);
+
+		BeanDefinitionBuilder gatewayProxyBuilder =
+				BeanDefinitionBuilder.rootBeanDefinition(GatewayProxyFactoryBean.class)
+						.addConstructorArgValue(serviceInterface);
+
+		if (hasDefaultHeaders || hasDefaultPayloadExpression) {
+			BeanDefinition methodMetadata = getMethodMetadataBeanDefinition(defaultPayloadExpression, defaultHeaders);
+
+			gatewayProxyBuilder.addPropertyValue("globalMethodMetadata", methodMetadata);
+		}
+
+		if (StringUtils.hasText(defaultRequestChannel)) {
+			gatewayProxyBuilder.addPropertyValue("defaultRequestChannelName", defaultRequestChannel);
+		}
+		if (StringUtils.hasText(defaultReplyChannel)) {
+			gatewayProxyBuilder.addPropertyValue("defaultReplyChannelName", defaultReplyChannel);
+		}
+		if (StringUtils.hasText(errorChannel)) {
+			gatewayProxyBuilder.addPropertyValue("errorChannelName", errorChannel);
+		}
+		if (asyncExecutor == null || AnnotationConstants.NULL.equals(asyncExecutor)) {
+			gatewayProxyBuilder.addPropertyValue("asyncExecutor", null);
+		}
+		else if (StringUtils.hasText(asyncExecutor)) {
+			gatewayProxyBuilder.addPropertyReference("asyncExecutor", asyncExecutor);
+		}
+		if (StringUtils.hasText(mapper)) {
+			gatewayProxyBuilder.addPropertyReference("mapper", mapper);
+		}
+		if (StringUtils.hasText(proxyDefaultMethods)) {
+			gatewayProxyBuilder.addPropertyValue(PROXY_DEFAULT_METHODS_ATTR, proxyDefaultMethods);
+		}
+
+		gatewayProxyBuilder.addPropertyValue("defaultRequestTimeoutExpressionString",
+				gatewayAttributes.get("defaultRequestTimeout"));
+		gatewayProxyBuilder.addPropertyValue("defaultReplyTimeoutExpressionString",
+				gatewayAttributes.get("defaultReplyTimeout"));
+		gatewayProxyBuilder.addPropertyValue("methodMetadataMap", gatewayAttributes.get("methods"));
+
+		String id = (String) gatewayAttributes.get("name");
+		if (!StringUtils.hasText(id)) {
+			BeanNameGenerator beanNameGenerator =
+					IntegrationConfigUtils.annotationBeanNameGenerator(registry);
+			id = beanNameGenerator.generateBeanName(new AnnotatedGenericBeanDefinition(serviceInterface), registry);
+		}
+
+		RootBeanDefinition beanDefinition = (RootBeanDefinition) gatewayProxyBuilder.getBeanDefinition();
+		beanDefinition.setTargetType(
+				ResolvableType.forClassWithGenerics(GatewayProxyFactoryBean.class, serviceInterface));
+		return new BeanDefinitionHolder(beanDefinition, id);
+	}
+
+	private static BeanDefinition getMethodMetadataBeanDefinition(String defaultPayloadExpression,
+			Map<String, Object>[] defaultHeaders) {
+
+		BeanDefinitionBuilder methodMetadataBuilder =
+				BeanDefinitionBuilder.genericBeanDefinition(GatewayMethodMetadata.class, GatewayMethodMetadata::new);
+
+		if (StringUtils.hasText(defaultPayloadExpression)) {
+			methodMetadataBuilder.addPropertyValue("payloadExpression",
+					BeanDefinitionBuilder.genericBeanDefinition(ExpressionFactoryBean.class)
+							.addConstructorArgValue(defaultPayloadExpression)
+							.getBeanDefinition());
+		}
+
+		if (!ObjectUtils.isEmpty(defaultHeaders)) {
+			Map<String, Object> headerExpressions = new ManagedMap<>();
+			for (Map<String, Object> header : defaultHeaders) {
+				String headerValue = (String) header.get("value");
+				String headerExpression = (String) header.get("expression");
+				boolean hasValue = StringUtils.hasText(headerValue);
+
+				if (hasValue == StringUtils.hasText(headerExpression)) {
+					throw new BeanDefinitionStoreException("exactly one of 'value' or 'expression' " +
+							"is required on a gateway's header.");
+				}
+
+				BeanDefinition expressionDef =
+						new RootBeanDefinition(hasValue ? LiteralExpression.class : ExpressionFactoryBean.class);
+				expressionDef.getConstructorArgumentValues()
+						.addGenericArgumentValue(hasValue ? headerValue : headerExpression);
+
+				headerExpressions.put((String) header.get("name"), expressionDef);
+			}
+			methodMetadataBuilder.addPropertyValue("headerExpressions", headerExpressions);
+		}
+		return methodMetadataBuilder.getBeanDefinition();
+	}
+
+	private static ConfigurableBeanFactory obtainBeanFactory(BeanDefinitionRegistry registry) {
+		if (registry instanceof ConfigurableBeanFactory) {
+			return (ConfigurableBeanFactory) registry;
+		}
+		else if (registry instanceof ConfigurableApplicationContext) {
+			return ((ConfigurableApplicationContext) registry).getBeanFactory();
+		}
+		throw new IllegalArgumentException("The provided 'BeanDefinitionRegistry' must be an instance " +
+				"of 'ConfigurableBeanFactory' or 'ConfigurableApplicationContext', but given is: "
+				+ registry.getClass());
+	}
+
+	private static Class<?> getServiceInterface(String serviceInterface, ConfigurableBeanFactory beanFactory) {
+		String actualServiceInterface = beanFactory.resolveEmbeddedValue(serviceInterface);
+		if (!StringUtils.hasText(actualServiceInterface)) {
+			return RequestReplyExchanger.class;
+		}
+		try {
+			return ClassUtils.forName(actualServiceInterface, beanFactory.getBeanClassLoader());
+		}
+		catch (ClassNotFoundException ex) {
+			throw new BeanDefinitionStoreException("Cannot parse class for service interface", ex);
 		}
 	}
